test(e2e): optimize e2e test structure
