+++ e2e/helpers/__init__.py
@@ -23,7 +23,18 @@ from .constants import (
     MCAST_ADDR,
     PROJECT_ROOT,
 )
-from .http import extract_catchup_source, http_get, http_request, stream_get, unix_http_get, unix_http_request
+from .config import build_config, build_single_service_config, write_temp_file
+from .http import (
+    assert_etag_cache_behavior,
+    extract_catchup_source,
+    get_header,
+    get_upstream_path,
+    http_get,
+    http_request,
+    stream_get,
+    unix_http_get,
+    unix_http_request,
+)
 from .mock_fcc import MockFCCServer
 from .mock_http import MockHTTPUpstream, MockHTTPUpstreamSilent
 from .mock_rtsp import (
@@ -62,10 +73,15 @@ __all__ = [
     "MockSTUNServer",
     "MulticastSender",
     "R2HProcess",
+    "assert_etag_cache_behavior",
+    "build_config",
+    "build_single_service_config",
     "extract_catchup_source",
     "find_free_port",
     "find_free_udp_port",
     "find_free_udp_port_pair",
+    "get_header",
+    "get_upstream_path",
     "http_get",
     "http_request",
     "ipv6_loopback_available",
@@ -76,4 +92,5 @@ __all__ = [
     "unix_http_request",
     "wait_for_port",
     "wait_for_unix_socket",
+    "write_temp_file",
 ]
